+++ Packages/webxr-interactions/CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An option to auto-load WebXRInputSystem on start in WebXRSettings.
 
 ### Changed
-- Minimum WebXR Input Profiles Loader version 0.6.0.
+- Minimum WebXR Input Profiles Loader version 0.6.2.
+
+### Fixed
+- SceneHitTest now updates the pose of the originTransform instead of WebXRManager transform.
 
 ## [0.20.0] - 2023-12-18
 ### Added